Add the initial `index.html`
This is Part A of Step 1 of the codelab "Build a Nearby Business Search
service wwith Google Maps Platform".

https://codelabs.developers.google.com/codelabs/google-maps-nearby-search-js/index.html

The content is just basic HTML. Most of the text is metadata and css in
the head. The body consists of just a single unused div, and it will be
the future home of the map.