Add initial javascript to load and display a map
This completes Part B of Step 1.

The `initMap` function sets some defaults, like being centered and
zoomed in on the opera house in Sydney. The function is then passed
in as the callback to the loading of API script.

When the page is loaded as is, the map will display Sydney but then
quickly is replaced by an error message directing you to check the
console. There you find a URL advising you that the current "key" is
invalid.

Minor note: there are inconsistencies in the indentation of the code
provided in the instructions. So when git diffing this file against
the copies in the stepN folders, it's good to add the `-b` or `-w`
flags so that whitespace differences are ignored.