Modify script to protect the API key
In order to avoid having the key be in the html, it's good to place the
key into a separate file. Now it can loaded and accessed via a
variable. This commit reuses the dummy key, but locally you can replace
it and then tell git to assume it is unchanged. (See first link.)

The catch is that src attribute of the script tag is outside of the
javascript itself, so it can't refer to any variables or concat strings
together.

A solution to this is to generate the script tag programatically and
then insert it in to the DOM. Not sure if there's any side effects to
this, but it's working on my machine.

- https://www.jimbobbennett.io/hiding-api-keys-from-git/
- https://stackoverflow.com/questions/14554928/add-javascript-variable-to-javascript-src