Add HTML5 geolocation
This completes Step 2.

The script now attempts to get the current lat+long from the browser,
and then initialize the map to be centered on that location.

The geolocation can fail, either because the browser lacks that
capability, or because the user did not grant permission to the page
to use it. So there's also a function to handle that, setting the
location to the original default and flash an appropriate message.