Add Places search request, create marker objects
This completes Part B of Step 3.

The code consists of a method to make the request, and a callback to
take the results and create markers on the map for them. The method
gets called in two places, the normal path and the error path for when
there's no geolocation.

The results exist in memory, but haven't been placed on the map yet,
so aren't visible. The `createMarkers` function is called but it
doesn't exist yet, resulting in an error message in the js console.