Add `createMarkers` function
This completes Part C of Step 3.

Basically a for-each loop over the results, where a marker is created
for each place and then the bounds object is updated to also include
the new place.

After the loop, the map adjusts to fit the updated bounds object.
This makes sure every place is visible within the view portal.