Add click listeners to markers on map
This completes Part B of Step 4.

The code is added inside the for-each loop during the marker creation.
A tiny anonymous function is wired to each marker, which will fire off
a request to the Places service to get details on the place. That
request also provides a callback function to handle the details when
they arrive.

Currently, there is no `showDetails` function. So clicking on a marker
fires off the API request, but then the js console spits out an error
after the request completes.