Add the `showDetails` function
This completes Part C of Step 4.

The function creates a standard informational ballon, filled in with
the name and rating for the place. The ballon then positions itself
based on the provided marker. Once that completes, the script tells
the previous ballon to hide, and then updates the currentInfoWindow
variable to point to the new ballon.

The `showPanel` function is not defined yet, so now a different error
message pops up in the console.

Small note: during the initialization, the currentInfoWindow variable
gets set to the balloon marking the location being searched from.
This also ensures that when the first marker is clicked then the close
attempt doesn't cause a null pointer exception.

Another note: the info windows are not saved or memoized. The upside
of this is that the results will always be fresh. But I think that
the old ballons are getting orphaned, and the page depends on the
garbage collector to free up the memory being used.