Add the `showPanel` function
This completes Part D of Step 4.

The code dynamically creates the content for the sidebar. It is kind
of long, because it creates and configures each DOM elements using js
commands instead of a template. On the other hand, maybe this is more
direct than generating HTML that then needs to be interpretted? Also,
there are a couple elements that are only conditionally being created.

The tutorial and code indicate that the panel is being closed and then
reopened, but in FF 70 on MacOS it's not visible. Too fast maybe?